commit our weight norm implementation
